app/requirements.txt
discord.py==2.3.2
cowsay==6.1